+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # @connect/connect-design-system
 
+## 1.23.0
+
+### Minor Changes
+
+- eb17ccc: - New Component:
+  - `<ButtonGroup />`: Button grouping with same styling as `<Button />`
+  - Updated Component:
+    - `<Button />`: Now using `ButtonBase` with MUI props
+  - Bug Fix: N/A
+  - Other: `ButtonBase` added to share MUI props between button components.
+
+## 1.22.0
+
+### Minor Changes
+
+- dbf3714: - Spacing values updated - Border radius values updated
+
+## 1.21.7
+
+### Patch Changes
+
+- ea8ebd6: Math icons added, icon viewbox updated to 20x20
+
+## 1.21.6
+
+### Patch Changes
+
+- b887933: Avatar - exposed ref
+
 ## 1.21.5
 
 ### Patch Changes