+++ src/components/ChangeLog.md
@@ -1,3 +1,83 @@
+### v1.23.0 Release Notice
+
+### Release date: March 12, 2025
+
+### Changelog:
+
+- New Component:
+  - `<ButtonGroup />`: Button grouping with same props/styling as `<Button />`
+- Updated Component:
+  - `<Button />`: Now using `ButtonBase` with MUI props
+- Bug Fix: N/A
+- Other: `ButtonBase` added to share props between button components.
+
+#### Repo:
+
+- Git Repo: https://scm.eng.hmhco.com/Core-Digital-Production/connect_system
+- Storybook: https://650211df9f7d78be97a6f39a-ggimoperby.chromatic.com/
+
+&nbsp;
+
+---
+
+### v1.22 Release Notice
+
+### Release date: March 12, 2025
+
+### Changelog:
+
+- Spacing values updated
+- Border radius values updated
+- Color sets added
+
+#### Repo:
+
+- Git Repo: https://scm.eng.hmhco.com/Core-Digital-Production/connect_system
+- Storybook: https://650211df9f7d78be97a6f39a-ggimoperby.chromatic.com/
+
+&nbsp;
+
+---
+
+### v1.21.7 Release Notice
+
+### Release date: March 10, 2025
+
+### Changelog:
+
+- Math icons added
+- Icons updated to 20x20 viewbox
+
+#### Repo:
+
+- Git Repo: https://scm.eng.hmhco.com/Core-Digital-Production/connect_system
+- Storybook: https://main--650211df9f7d78be97a6f39a.chromatic.com
+
+&nbsp;
+
+---
+
+### v1.21.6 Release Notice
+
+### Release date: March 10, 2025
+
+### Changelog:
+
+- New Component: N/A
+- Updated Component:
+  - `<Avatar />`: Exposed ref
+- Bug Fix: N/A
+- Other: N/A
+
+#### Repo:
+
+- Git Repo: https://scm.eng.hmhco.com/Core-Digital-Production/connect_system
+- Storybook: https://650211df9f7d78be97a6f39a-ggimoperby.chromatic.com/
+
+&nbsp;
+
+---
+
 ### v1.21.5 Release Notice
 
 ### Release date: February 28, 2025